% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockandstone.R
\name{play_voiceline}
\alias{play_voiceline}
\title{Play voicelines}
\usage{
play_voiceline(file_ogg)
}
\arguments{
\item{file_ogg}{The .ogg file path}
}
\value{
\code{NULL}
}
\description{
Play voicelines
}
