% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockandstone.R
\name{rockandstone}
\alias{rockandstone}
\title{Play a random voiceline}
\usage{
rockandstone(category = NULL, sentiment = NULL, nsfw = FALSE)
}
\arguments{
\item{category}{An optional character vector specifying the voiceline
categories to include for playback. See \code{voiceline_categories()} for
options.}

\item{sentiment}{An optional character vector specifying the voiceline
sentiments to include for playback. Options are "positive", "negative",
and "neutral".}

\item{nsfw}{A logical specifying whether to include "Not Safe For Work"
voicelines that include profanity. Default is \code{FALSE}.}
}
\value{
\code{NULL}
}
\description{
\code{rockandstone()} plays a random voiceline. This is useful, for example, to
to get notified when a script is finished.
}
\note{
\href{https://deeprockgalactic.wiki.gg/wiki/Voicelines}{Voicelines} by
\href{https://ghostship.dk}{Ghost Ship Games} are licensed under
\href{https://creativecommons.org/licenses/by-nc-sa/3.0/}{CC BY-NC-SA 3.0}
}
\examples{

# By default the voiceline pool includes all "Safe For Work" voicelines
rockandstone()

# "Not Safe For Work Voicelines" (with profanity) are opt-in
rockandstone(nsfw = TRUE)

# The voiceline pool can be reduced to certain categories and/or sentiments
# to suit your use case
rockandstone(category = "salute")
rockandstone(sentiment = "negative")
}
