% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockandstone.R
\name{select_voiceline}
\alias{select_voiceline}
\title{Select a random voiceline}
\usage{
select_voiceline(category = NULL, sentiment = NULL, nsfw = FALSE)
}
\description{
Same arguments as \code{rockandstone()}.
}
\keyword{internal}
